Puppet::Resource::Type (evaluate_produces): new method
Evaluate and instantiate produced resources, then tag them with
producer:<environment>. This makes it possibly to find the produced
resource as copies of this resource will wind up in the catalogs of nodes
that use it.

One consequence of this change is that capability resources are evaluated
in their own scope, separate from the body of the define that produces them
(since we aren't actually evaluating that define during environment catalog
compilation). During node compilation, that means the same scope will be
created twice, and therefore any variables set during evaluation of the
`produces` expression won't be visible to the body of the definition. In
practice this shouldn't be a problem, because the only way a variable could
be defined at that time is through a strange function call.