A Puppet face to extract the order in which Puppet needs to run
This is the beginning of an environment/application graph compiler, for
simplicity all contained in one file for now

To try this out, run

    bundle exec ./bin/puppet app example.pp

The code is somewhat bolted onto the side of the existing compiler, and
therefore doesn't support all Puppet language features. It does work with
the examples from
https://github.com/puppetlabs/xnode-spec/tree/master/examples (except for
40-riak.pp)

There are lots of things that need to be fixed about this code, but it does
the salient bits of analyzing an application instance and determining in
which order Puppet needs to be run. Things that need to be changed are

  * move some of the code out of lib/puppet/application/app.rb into more
    logical places
  * figure out how to integrate this better with the node catalog compiler
  * hook this up to something that actually runs Puppet (ssh, mco, pegasus, ...)
  * need code that will pull component definitions into the catalog of nodes
    when the node requests its catalog

None of these are major issues - it's pretty clear how to do that given
that we can extract application instances from manifests (which this code
does)