Introduce an EnvironmentCompiler
This compiler subclasses the ordinary node compiler, and is used for
compiling the environment catalog. It differs from the regular compiler
by only evaluating enough to find the components and capability
resources. Any other resources will cause a failure.

The $manifest directory is parsed, assuming all application declarations
happen at top-level. Then the main class (which contains the
applications) is evaluated, then the applications are evaluated to find
components, then the `produces` statements (only) are evaluated to build
the capability resources.

One consequence of this change is that capability resources are
evaluated in their own scope, separate from the body of the define that
produces them (since we aren't actually evaluating that define during
environment catalog compiler). During node compilation, that means the
same scope will be created twice, and therefore any variables set during
evaluation of the `produces` expression won't be visible to the body of
the definition. In practice this shouldn't be a problem, because the
only way a variable could be defined at that time is through a strange
function call.

This compiler is used by the /v3/environment/<name> route, which consumes
the catalog and walks it to build the appropriate return structure. If the
environment <name> does not exist, the handler returns a 404 (see
APPMGMT-144)

Copying the regular compiler allows features like autoloading to work,
so we can use a proper module layout.

Fixes APPMGMT-53, APPMGMT-69, APPMGMT-69, APPMGMT-144